added test and fixed the playmove usecase
