+++ src/main/java/scrabble_controller/ScrabbleGameController.java
@@ -20,6 +20,7 @@ import usecases.usecase_implementations.PlayMove;
 import usecases.usecase_interfaces.*;
 import java.util.ArrayList;
 import gui.View;
+import java.util.List;
 import usecases.usecase_implementations.ScrabbleDictionary;
 
 /**
@@ -107,11 +108,33 @@ public class ScrabbleGameController{
     
 
     /**
-     * This method is responsible for calling the playmove usecase. 
+     * This method is responsible for calling . 
      */
     public void playMove() {
-        ((PlayMoveUsecase)playMove).playMove(handManager, boardManager, playerManager, turnManager, gameScorer, game,
-                scrabbleDictionary, this);
+        GameBoard prevBoard = boardManager.getPrevBoard();
+        List<List<List<Integer>>> words = ((PlaceWordUsecase) boardManager).checkWord(game, scrabbleDictionary, prevBoard);
+
+        //boardmanager checkword returns list of coordinates and list of letters used by the player
+
+        if(!words.isEmpty()) {
+            System.out.println("valid word");
+            // ScoringSystem
+            int score = ((CalculateWordScoreUsecase) gameScorer).calculateMultiWordScore(game, words);
+            // calculate the total score of all the words found
+            ((UpdateScoreUsecase) playerManager).updateScoreForCurrentPlayer(game.getCurrentPlayer().getScore() + score, game);
+            // place word usecase
+
+
+            ((IncrementTurnUsecase) turnManager).incrementTurn(game);
+
+            ((FillHandUsecase)handManager).fillHand(game);// fill the next player's hand
+            
+            ((ResetMoveUsecase)boardManager).clearMoves();
+        }
+        else {
+            System.out.println("not valid");
+            resetMove();
+        }
         saveGameToFile();
         view.updateView(game);
     }
