+++ src/main/java/usecases/usecase_implementations/BoardManager.java
@@ -1,6 +1,6 @@
 package usecases.usecase_implementations;
 
-import entities.MoveInfo;
+//import entities.MoveInfo;
 import entities.Cell;
 import entities.Game;
 import entities.GameBoard;
@@ -59,12 +59,9 @@ public class BoardManager implements PlaceTileUsecase, PlaceWordUsecase, ResetMo
         ArrayList<List<Integer>> move_list = new ArrayList<>();
         createListOfCoordinates(move_list);
         TileChecker validate_word = new TileChecker();
-        if (prevBoard.isEmpty()) { // check if it's first turn of thr game
+        if (prevBoard.isEmpty()) { // check if it's first turn of the game
             if (checkFirstTurnCondition()) { // check if the word is on center of board
                 ArrayList<List<List<Integer>>> first_word_list = validate_word.validateMove(move_list, game.getGameBoard(), scrabbleDictionary, prevBoard);
-                if (first_word_list.isEmpty()) {
-                    resetMoves(game); // change board back to previous state if no valid words.
-                }
                 return first_word_list; // return list of coordinates of new word on board
             }
             else {
@@ -73,9 +70,7 @@ public class BoardManager implements PlaceTileUsecase, PlaceWordUsecase, ResetMo
         }
         else {
             ArrayList<List<List<Integer>>> word_list = validate_word.validateMove(move_list, game.getGameBoard(), scrabbleDictionary, prevBoard);
-            if (word_list.isEmpty()) {
-                resetMoves(game); // change board back to previous state if no valid words.
-            }
+            clearMoves();
             return word_list; // return list of valid words that can be made from the moves
         }
     }
@@ -118,6 +113,7 @@ public class BoardManager implements PlaceTileUsecase, PlaceWordUsecase, ResetMo
     /**
      * This method returns the variable previous_board from BoardManager class.
      * @return previous_board, a GameBoard object with the previous moves.
+     *
      */
     public GameBoard getPrevBoard(){
         return this.previous_board;
