+++ src/test/java/controllers/ScrabbleGameControllerTest.java
@@ -1,0 +1,19 @@
+package controllers;
+
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
+/**
+ *
+ * @author jingw
+ */
+public class ScrabbleGameControllerTest {
+    
+    
+    
+    
+    
+    
+}
