+++ src/test/java/use_case_implementations/PlayMoveTest.java
@@ -2,16 +2,12 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
+package use_case_implementations;
 
 /**
  *
  * @author jingw
  */
-public class ScrabbleGameControllerTest {
-    
-    
-    
-    
-    
+public class PlayMoveTest {
     
 }
